+++ src/libs/pestpp_common/SVDSolver.cpp
@@ -28,6 +28,7 @@
 #include <map>
 #include <algorithm>
 #include <sstream>
+#include <cmath>
 #include "SVDSolver.h"
 #include "RunManagerAbstract.h"
 #include "QSqrtMatrix.h"
@@ -1211,12 +1212,12 @@ ModelRun SVDSolver::iteration_upgrd(RunManagerAbstract &run_manager, Termination
             
 			// Determine the number of scales to use
             int num_lrun_level = (lmrun - maxitn) / maxitn + 1;
-            if (num_lrun_level > lambda_scale_vec.size()) {
-                num_lrun_level = lambda_scale_vec.size();
+            if (num_lrun_level > static_cast<int>(lambda_scale_vec.size())) {
+                num_lrun_level = static_cast<int>(lambda_scale_vec.size());
             }
 
             // Slice the line search factor vector in place
-            if (num_lrun_level < lambda_scale_vec.size()) {
+            if (num_lrun_level < static_cast<int>(lambda_scale_vec.size())) {
                 lambda_scale_vec.resize(num_lrun_level);
             }
 			
@@ -1225,7 +1226,7 @@ ModelRun SVDSolver::iteration_upgrd(RunManagerAbstract &run_manager, Termination
 			const double def_lowest_lambda_fac = 1.0e-14;
 			const double def_highest_lambda_fac = 1.0e14;
 			double phi_initial = base_run.get_phi(*regul_scheme_ptr);
-			double num_nonzero_obs = obs_names_vec.size();
+			int num_nonzero_obs = static_cast<int>(obs_names_vec.size());
 			// this is the critical difference between GLM lambdas and HP lambdas
 			// HP lambdas are a factor of the current phi
 			double lambda_mid = phi_initial / static_cast<double>(num_nonzero_obs);
